+++ docs/changelog.rst
@@ -18,6 +18,9 @@ version **0.2.4** (Not yet released)
 * return the result of the condition checked by ``wait_for`` and
   ``wait_for_and_refresh``.
 * start documenting the debian/ubuntu packaging process.
+* removed HTML results reports.
+* switched to `junitxml` dependency for XML report generation.
+
 
 version **0.2.3** (2013 Apr 17)
 *******************************